Produce policy JSON rather than policy ARNs
We currently produce a list of IAM policy ARNs which can be attached to
a role to provide permissions to access an addon's resources. This
produces an excess of IAM policies which will usually only be applied to
a single role. It also runs the risk of exhausting the quota of 10 IAM
policies per role, particularly when policies for access secrets and
keys is added.

This updates the addon modules to produce policy JSON rather than
actually creating policies. This JSON can then be aggregated together to
create a single policy for accessing the combined resources of the
addons a role will use.

In order to make this easier to use, two other changes are included:

- Policies will now ensure they use unique SIDs for each addon so that
  policies can be aggregated together by SID
- A new module is included to create a policy from aggregated JSON and
  attach it to a role