Split read, admin trust principals
We currently accept a single principal which will be trusted in the
resource policy for secrets and keys. This is problematic when we want
different accounts or other principals to be trusted.

This introduces separate admin and read principals. It also accepts a
list so that more than one principal may be introduced for each role.